In fact, the `savingsTargetEnabled` feature flag should be global, even though it's reported in the `Account` response for convenience, so I have removed the `FeatureFlagsRepo` and shifted the flag to the application config.
I have also taken the opportunity to split up the `HelpToSaveControllerSpec` into three separate spec classes as the source file was getting way too big.
