+++ conf/application.conf
@@ -189,6 +189,7 @@ router {
 helpToSave {
   supportFormEnabled = true
   inAppPaymentsEnabled = false
+  savingsTargetsEnabled = true
   infoUrl = "https://www.gov.uk/get-help-savings-low-income"
   invitationUrl = "http://localhost:8249/mobile-help-to-save"
   accessAccountUrl = "http://localhost:8249/mobile-help-to-save/access-account"
